Arbitrary-length simd

+++ src/raze/math/CountTrailingZeros.h
@@ -38,8 +38,8 @@ __RAZE_MATH_NAMESPACE_BEGIN
 #  endif // defined(raze_cpp_clang) || defined(raze_cpp_gnu) || defined(raze_cpp_msvc)
 #endif // defined (raze_processor_x86)
 
-constexpr raze_always_inline int __bit_hacks_ctz_u32(uint32 __value) noexcept {
-    auto __result   = uint32(32);
+constexpr raze_always_inline i32 __bit_hacks_ctz_u32(u32 __value) noexcept {
+    auto __result   = u32(32);
     __value         &= -signed(__value);
 
     if (__value)
@@ -64,19 +64,19 @@ constexpr raze_always_inline int __bit_hacks_ctz_u32(uint32 __value) noexcept {
 }
 
 template <std::unsigned_integral _IntegralType_>
-constexpr raze_always_inline int __bit_hacks_ctz(_IntegralType_ __value) noexcept {
+constexpr raze_always_inline i32 __bit_hacks_ctz(_IntegralType_ __value) noexcept {
     if constexpr (sizeof(_IntegralType_) == 8) {
-        const auto __low = static_cast<uint32>(__value);
+        const auto __low = static_cast<u32>(__value);
         return __low
             ? __bit_hacks_ctz_u32(__low)
-            : 32 + __bit_hacks_ctz_u32(static_cast<uint32>(__value >> 32));
+            : 32 + __bit_hacks_ctz_u32(static_cast<u32>(__value >> 32));
     }
     else if constexpr (sizeof(_IntegralType_) == 4) {
-        return __bit_hacks_ctz_u32(static_cast<uint32>(__value));
+        return __bit_hacks_ctz_u32(static_cast<u32>(__value));
     }
     else if constexpr (sizeof(_IntegralType_) == 2) {
-        auto __result   = uint32(16);
-        __value         &= uint16(-signed(__value));
+        auto __result   = u32(16);
+        __value         &= u16(-signed(__value));
         
         if (__value)
             --__result;
@@ -96,8 +96,8 @@ constexpr raze_always_inline int __bit_hacks_ctz(_IntegralType_ __value) noexcep
         return __result;
     }
     else if constexpr (sizeof(_IntegralType_) == 1) {
-        auto __result   = uint32(8);
-        __value         &= uint8(-signed(__value));
+        auto __result   = u32(8);
+        __value         &= u8(-signed(__value));
 
         if (__value)
             --__result;
@@ -119,7 +119,7 @@ constexpr raze_always_inline int __bit_hacks_ctz(_IntegralType_ __value) noexcep
 #if defined (raze_processor_x86)
 
 template <std::unsigned_integral _IntegralType_>
-raze_always_inline int __bsf_ctz(_IntegralType_ __value) noexcept {
+raze_always_inline i32 __bsf_ctz(_IntegralType_ __value) noexcept {
     constexpr auto __digits = std::numeric_limits<_IntegralType_>::digits;
     constexpr auto __max    = std::numeric_limits<_IntegralType_>::max();
 
@@ -134,7 +134,7 @@ raze_always_inline int __bsf_ctz(_IntegralType_ __value) noexcept {
             return __digits;
     }
     else if constexpr (__digits < 32) {
-        if (!_BitScanForward(&__index, static_cast<uint32>(__value | ~__max)))
+        if (!_BitScanForward(&__index, static_cast<u32>(__value | ~__max)))
             return __digits;
     }
 
@@ -142,22 +142,22 @@ raze_always_inline int __bsf_ctz(_IntegralType_ __value) noexcept {
 }
 
 template <std::unsigned_integral _IntegralType_>
-raze_always_inline int __tzcnt_ctz(_IntegralType_ __value) noexcept {
+raze_always_inline i32 __tzcnt_ctz(_IntegralType_ __value) noexcept {
     constexpr auto __digits = std::numeric_limits<_IntegralType_>::digits;
     constexpr auto __max    = std::numeric_limits<_IntegralType_>::max();
 
     if      constexpr (__digits == 64)
-        return static_cast<int>(__raze_tzcnt_u64(__value));
+        return __raze_tzcnt_u64(__value);
     else if constexpr (__digits == 32)
-        return static_cast<int>(__raze_tzcnt_u32(__value));
+        return __raze_tzcnt_u32(__value);
     else if constexpr (__digits < 32)
-        return static_cast<int>(__raze_tzcnt_u32(static_cast<uint32>(__value | ~__max)));
+        return __raze_tzcnt_u32(static_cast<u32>(__value | ~__max));
 }
 
 #endif // defined(raze_processor_x86)
 
 template <std::unsigned_integral _IntegralType_>
-constexpr raze_always_inline int __count_trailing_zero_bits(_IntegralType_ __value) noexcept {
+constexpr raze_always_inline i32 __count_trailing_zero_bits(_IntegralType_ __value) noexcept {
     static_assert(std::is_unsigned_v<_IntegralType_>);
 
 #if defined(raze_processor_x86) && !defined(raze_processor_arm)
@@ -177,7 +177,7 @@ constexpr raze_always_inline int __count_trailing_zero_bits(_IntegralType_ __val
 template <sizetype _Bits_> 
 struct __ctz_n_bits_implementation {
     template <std::unsigned_integral _IntegralType_>
-    constexpr raze_always_inline int operator()(_IntegralType_ __value) const noexcept {
+    constexpr raze_always_inline i32 operator()(_IntegralType_ __value) const noexcept {
         constexpr auto __max_for_n_bits = _IntegralType_(((_IntegralType_(1) << _Bits_) - 1));
         constexpr auto __mask_size = (_Bits_ / 8) > 1 ? (_Bits_ / 8) : 1;
 
@@ -190,9 +190,9 @@ struct __ctz_n_bits_implementation {
 template <>
 struct __ctz_n_bits_implementation<2> {
     template <std::unsigned_integral _IntegralType_>
-    constexpr raze_always_inline int operator()(_IntegralType_ __value) const noexcept {
-        auto __result   = uint32(2);
-        __value         &= uint8(-signed(__value));
+    constexpr raze_always_inline i32 operator()(_IntegralType_ __value) const noexcept {
+        auto __result   = u32(2);
+        __value         &= u8(-signed(__value));
 
         if (__value)
             --__result;
@@ -208,9 +208,9 @@ struct __ctz_n_bits_implementation<2> {
 template <> 
 struct __ctz_n_bits_implementation<4> {
     template <std::unsigned_integral _IntegralType_>
-    constexpr raze_always_inline int operator()(_IntegralType_ __value) const noexcept {
-        auto __result   = uint32(4);
-        __value         &= uint8(-signed(__value));
+    constexpr raze_always_inline i32 operator()(_IntegralType_ __value) const noexcept {
+        auto __result   = u32(4);
+        __value         &= u8(-signed(__value));
 
         if (__value)
             --__result;
@@ -228,7 +228,7 @@ struct __ctz_n_bits_implementation<4> {
 template <
     sizetype _Bits_, 
     std::unsigned_integral _IntegralType_>
-constexpr raze_always_inline int __ctz_n_bits(_IntegralType_ __value) noexcept {
+constexpr raze_always_inline i32 __ctz_n_bits(_IntegralType_ __value) noexcept {
     static_assert(_Bits_ <= 64);
     static_assert(raze_sizeof_in_bits(_IntegralType_) >= _Bits_);
